AGENT-277: covered_remove_log_path, added stale request test. start from checkpoint test status generation test. logs_initial_position test schedule_pending_log_path_for_removal

+++ scalyr_agent/tests/copying_manager_test.py
@@ -245,6 +245,29 @@ class DynamicLogPathTest(ScalyrTestCase):
         matchers = self._manager.log_matchers
         self.assertEquals( 0, len(matchers) )
 
+    def test_schedule_pending_log_path_for_removal(self):
+        config = {}
+        self.create_copying_manager( config )
+        self.fake_scan()
+
+        path = os.path.join( self._log_dir, "newlog.log" )
+        self.append_log_lines( path, "line1\n" )
+
+        log_config = {
+            "path": path
+        }
+
+        self._manager.add_log_config('unittest', log_config)
+        self._manager.schedule_log_path_for_removal('unittest', path)
+
+        self.fake_scan()
+        self.fake_scan()
+        self.fake_scan()
+
+        matchers = self._manager.log_matchers
+
+        self.assertEquals( 0, len(matchers) )
+
     def test_schedule_log_path_for_removal_different_monitor(self):
         config = {}
         self.create_copying_manager( config )
@@ -270,25 +293,38 @@ class DynamicLogPathTest(ScalyrTestCase):
 
     def test_remove_log_path(self):
         config = {}
-        self.create_copying_manager( config )
+        self.create_copying_manager(config)
         self.fake_scan()
 
-        path = os.path.join( self._log_dir, "newlog.log" )
-        self.append_log_lines( path, "line1\n" )
+        path = os.path.join(self._log_dir, "newlog.log")
+        self.append_log_lines(path, "line1\n")
 
         log_config = {
             "path": path
         }
 
-        self._manager.add_log_config( 'unittest', log_config )
+        path2 = os.path.join(self._log_dir, "newlog2.log")
+        self.append_log_lines(path2, "line1\n")
+
+        log_config2 = {
+            "path": path2
+        }
+
+        self._manager.add_log_config('unittest', log_config)
+        self._manager.add_log_config('unittest', log_config2)
         self.fake_scan()
         matchers = self._manager.log_matchers
-        self.assertEquals( 1, len(matchers) )
-        self.assertEquals( path, matchers[0].log_path )
+        self.assertEquals(2, len(matchers))
+        self.assertEquals(path, matchers[0].log_path)
+        self.assertEquals(path2, matchers[1].log_path)
 
-        self._manager.remove_log_path( 'unittest', path )
+        self._manager.remove_log_path('unittest', path)
         matchers = self._manager.log_matchers
-        self.assertEquals( 0, len(matchers) )
+        self.assertEquals(1, len(matchers))
+
+        self._manager.remove_log_path('unittest', path2)
+        matchers = self._manager.log_matchers
+        self.assertEquals(0, len(matchers))
 
     def test_add_after_remove(self):
         config = {}
@@ -588,6 +624,7 @@ class CopyingManagerEnd2EndTest(ScalyrTestCase):
     def setUp(self):
         super(CopyingManagerEnd2EndTest, self).setUp()
         self._controller = None
+        self._config = None
 
     def tearDown(self):
         if self._controller is not None:
@@ -784,6 +821,182 @@ class CopyingManagerEnd2EndTest(ScalyrTestCase):
 
         responder_callback('success')
 
+    def test_start_from_full_checkpoint(self):
+        controller = self.__create_test_instance()
+        previous_root_dir = os.path.dirname(self.__test_log_file)
+
+        self.__append_log_lines('First line', 'Second line')
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+
+        self.assertEquals(2, len(lines))
+        self.assertEquals('First line', lines[0])
+        self.assertEquals('Second line', lines[1])
+
+        # stop thread on manager to write checkouts to file.
+        controller.stop()
+
+        # write some new lines to log.
+        self.__append_log_lines('Third line', 'Fourth line')
+
+        # create new copying manager, but passing previous directory with same log and checkouts.
+        controller = self.__create_test_instance(root_dir=previous_root_dir, auto_start=False)
+
+        self._manager.start_manager(dict(fake_client=True))
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+
+        # thread should continue from saved checkpoint
+        self.assertEquals(2, len(lines))
+        self.assertEquals('Third line', lines[0])
+        self.assertEquals('Fourth line', lines[1])
+
+        # stopping one more time, but now emulating that checkpoint files are stale.
+        controller.stop()
+
+        self.__append_log_lines('Fifth line', 'Sixth line')
+
+        def shift_time_in_checkpoint_file(name):
+
+            with open(os.path.join(self._config.agent_data_path, name), 'r') as fp:
+                data = json_lib.parse(fp.read())
+
+            data['time'] -= self._config.max_allowed_checkpoint_age + 1
+
+            with open(os.path.join(self._config.agent_data_path, name), 'w') as fp:
+                fp.write(json_lib.serialize(data))
+
+        # shift time on checkpoint files in the past, to make them stale.
+        shift_time_in_checkpoint_file('checkpoints.json')
+        shift_time_in_checkpoint_file('active-checkpoints.json')
+
+        # start manager.
+        controller = self.__create_test_instance(root_dir=previous_root_dir, auto_start=False)
+        self._manager.start_manager(dict(fake_client=True))
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+        self.assertEquals(0, len(lines))
+
+    def test_start_from_active_checkpoint(self):
+        controller = self.__create_test_instance()
+        previous_root_dir = os.path.dirname(self.__test_log_file)
+
+        self.__append_log_lines('First line', 'Second line')
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+
+        self.assertEquals(2, len(lines))
+        self.assertEquals('First line', lines[0])
+        self.assertEquals('Second line', lines[1])
+
+        controller.stop()
+
+        with open(os.path.join(self._config.agent_data_path, 'active-checkpoints.json'), 'r') as fp:
+            data = json_lib.parse(fp.read())
+
+        data['time'] += 100
+
+        with open(os.path.join(self._config.agent_data_path, 'active-checkpoints.json'), 'w') as fp:
+            fp.write(json_lib.serialize(data))
+
+        self.__append_log_lines('Third line', 'Fourth line')
+
+        controller = self.__create_test_instance(root_dir=previous_root_dir, auto_start=False)
+        self._manager.start_manager(dict(fake_client=True))
+
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+        self.assertEquals(2, len(lines))
+        self.assertEquals('Third line', lines[0])
+        self.assertEquals('Fourth line', lines[1])
+
+    def test_start_without_active_checkpoint(self):
+        controller = self.__create_test_instance()
+        previous_root_dir = os.path.dirname(self.__test_log_file)
+
+        self.__append_log_lines('First line', 'Second line')
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+
+        self.assertEquals(2, len(lines))
+        self.assertEquals('First line', lines[0])
+        self.assertEquals('Second line', lines[1])
+        controller.stop()
+
+        self.__append_log_lines('Third line', 'Fourth line')
+
+        os.remove(os.path.join(self._config.agent_data_path, 'active-checkpoints.json'))
+
+        controller = self.__create_test_instance(root_dir=previous_root_dir, auto_start=False)
+        self._manager.start_manager(dict(fake_client=True))
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+        self.assertEquals(2, len(lines))
+        self.assertEquals('Third line', lines[0])
+        self.assertEquals('Fourth line', lines[1])
+
+    def test_stale_request(self):
+        controller = self.__create_test_instance()
+        self.__append_log_lines('First line', 'Second line')
+        (request, responder_callback) = controller.wait_for_rpc()
+
+        lines = self.__extract_lines(request)
+        self.assertEquals(2, len(lines))
+        self.assertEquals('First line', lines[0])
+        self.assertEquals('Second line', lines[1])
+
+        from scalyr_agent import copying_manager
+
+        # backup original 'time' module
+        orig_time = copying_manager.time
+
+        class _time_mock(object):
+            # This dummy 'time()' should be called on new copying thread iteration
+            # to emulate huge gap between last request.
+            def time(self):
+                # return time.time() + 'CopyingManager.__config.max_retry_time'
+                res = orig_time.time() + 901
+                # restore original 'time' module
+                copying_manager.time = orig_time
+                return res
+
+        # repalce time module with dummy time time object.
+        copying_manager.time = _time_mock()
+
+        responder_callback('error')
+
+        (request, responder_callback) = controller.wait_for_rpc()
+
+        lines = self.__extract_lines(request)
+        self.assertEquals(0, len(lines))
+
+    def test_generate_status(self):
+        controller = self.__create_test_instance()
+
+        self.__append_log_lines('First line', 'Second line')
+        (request, responder_callback) = controller.wait_for_rpc()
+        lines = self.__extract_lines(request)
+
+        self.assertEquals(2, len(lines))
+        self.assertEquals('First line', lines[0])
+        self.assertEquals('Second line', lines[1])
+
+        status = self._manager.generate_status()
+
+        self.assertEquals(2, len(status.log_matchers))
+
+    def test_logs_initial_positions(self):
+        controller = self.__create_test_instance(auto_start=False)
+        self._manager.start_manager(dict(fake_client=True), logs_initial_positions={self.__test_log_file: 5*2})
+
+        self.__append_log_lines(*[str(i) for i in range(10)])
+
+        request , cb = controller.wait_for_rpc()
+
+        lines = self.__extract_lines(request)
+
+        self.assertEquals(['5', '6', '7', '8', '9'], lines)
+
     def __extract_lines(self, request):
         parsed_request = json_lib.parse(request.get_payload())
 
@@ -801,23 +1014,30 @@ class CopyingManagerEnd2EndTest(ScalyrTestCase):
     def __was_pipelined(self, request):
         return 'pipelined=1.0' in request.get_timing_data()
 
-    def __create_test_instance(self, use_pipelining=False):
-        tmp_dir = tempfile.mkdtemp()
-        config_dir = os.path.join(tmp_dir, 'config')
-        data_dir = os.path.join(tmp_dir, 'data')
-        log_dir = os.path.join(tmp_dir, 'log')
+    def __create_test_instance(self, use_pipelining=False, root_dir=None, auto_start=True):
+        if root_dir is None:
+            root_dir = tempfile.mkdtemp()
+        config_dir = os.path.join(root_dir, 'config')
+        data_dir = os.path.join(root_dir, 'data')
+        log_dir = os.path.join(root_dir, 'log')
 
-        os.mkdir(data_dir)
-        os.mkdir(config_dir)
-        os.mkdir(log_dir)
+        if not os.path.exists(data_dir):
+            os.mkdir(data_dir)
+        if not os.path.exists(config_dir):
+            os.mkdir(config_dir)
+        if not os.path.exists(log_dir):
+            os.mkdir(log_dir)
 
-        self.__test_log_file = os.path.join(tmp_dir, 'test.log')
-        fp = open(self.__test_log_file, 'w')
-        fp.close()
+        self.__test_log_file = os.path.join(root_dir, 'test.log')
+
+        if not os.path.exists(self.__test_log_file):
+            fp = open(self.__test_log_file, 'w')
+            fp.close()
 
         config_file = os.path.join(config_dir, 'agentConfig.json')
         config_fragments_dir = os.path.join(config_dir, 'configs.d')
-        os.makedirs(config_fragments_dir)
+        if not os.path.exists(config_fragments_dir):
+            os.makedirs(config_fragments_dir)
 
         logs_json_array = JsonArray()
         logs_json_array.add(JsonObject(path=self.__test_log_file))
@@ -826,18 +1046,22 @@ class CopyingManagerEnd2EndTest(ScalyrTestCase):
         if use_pipelining:
             pipeline_threshold = 0.0
 
-        fp = open(config_file, 'w')
-        fp.write(json_lib.serialize(JsonObject(api_key='fake', logs=logs_json_array,
-                                               pipeline_threshold=pipeline_threshold)))
-        fp.close()
+        if not os.path.exists(config_file):
+            fp = open(config_file, 'w')
+            fp.write(json_lib.serialize(JsonObject(api_key='fake', logs=logs_json_array,
+                                                   pipeline_threshold=pipeline_threshold)))
+            fp.close()
 
         default_paths = DefaultPaths(log_dir, config_file, data_dir)
 
         config = Configuration(config_file, default_paths, None)
         config.parse()
 
+        self._config = config
+
         # noinspection PyTypeChecker
-        self._controller = TestableCopyingManager(config, []).controller
+        self._manager = TestableCopyingManager(config, [], auto_start=auto_start)
+        self._controller = self._manager.controller
         return self._controller
 
     def __append_log_lines(self, *args):
@@ -866,7 +1090,7 @@ class TestableCopyingManager(CopyingManager):
     # To prevent tests from hanging indefinitely, wait a maximum amount of time before giving up on some test condition.
     WAIT_TIMEOUT = 5.0
 
-    def __init__(self, configuration, monitors):
+    def __init__(self, configuration, monitors, auto_start=True):
         CopyingManager.__init__(self, configuration, monitors)
         # Approach:  We will override key methods of CopyingManager, blocking them from returning until the controller
         # tells it to proceed.  This allows us to then do things like write new log lines while the CopyingManager is
@@ -895,7 +1119,7 @@ class TestableCopyingManager(CopyingManager):
         # Protected by __test_state_cv.  The status message to return for the next call to ``_send_events``.
         self.__pending_response = None
 
-        self.__controller = TestableCopyingManager.TestController(self)
+        self.__controller = TestableCopyingManager.TestController(self, auto_start=auto_start)
 
     @property
     def controller(self):
@@ -1093,12 +1317,14 @@ class TestableCopyingManager(CopyingManager):
 
         Its main role is to tell the manager thread when to unblock and how far to run.
         """
-        def __init__(self, copying_manager):
+
+        def __init__(self, copying_manager, auto_start=True):
             self.__copying_manager = copying_manager
-            copying_manager.start_manager(dict(fake_client=True))
-            # To do a proper initialization where the copying manager has scanned the current log file and is ready
-            # for the next loop, we let it go all the way through the loop once and wait in the sleeping state.
-            copying_manager.run_and_stop_at(TestableCopyingManager.SLEEPING)
+            if auto_start:
+                copying_manager.start_manager(dict(fake_client=True))
+                # To do a proper initialization where the copying manager has scanned the current log file and is ready
+                # for the next loop, we let it go all the way through the loop once and wait in the sleeping state.
+                copying_manager.run_and_stop_at(TestableCopyingManager.SLEEPING)
 
         def perform_scan(self):
             """Tells the CopyingManager thread to go through the process loop until far enough where it has performed
